V10: working

+++ sound/soc/codecs/hi6402.c
@@ -245,7 +245,7 @@ void hi6402_set_rec_pll(struct snd_soc_codec *codec,unsigned int offset,unsigned
 	if(true == state)
 		rec_state = m_rec_state | (1 << offset);
 	else
-		rec_state = m_rec_state & (~(1 << offset));
+		rec_state = m_rec_state & (~(1ul << offset));
 }
 
 int hi6402_mad_power_mode_event(struct snd_soc_dapm_widget *w,
@@ -571,6 +571,7 @@ int hi6402_slimbus_port1415_power_mode_event(struct snd_soc_dapm_widget *w,
 	return ret;
 }
 
+/* start: add for dapm to solution 3mic conflict */
 int hi6402_s1_interface_clk_power_mode_event(struct snd_soc_dapm_widget *w,
 				struct snd_kcontrol *kcontrol, int event)
 {
@@ -901,7 +902,7 @@ static void hi6402_hp_power_on(struct snd_soc_codec *codec)
 	unsigned int hsl_pga_value = hsl_reg_value & 0x3F;
 	unsigned int hsr_pga_value = hsr_reg_value & 0x3F;
 	unsigned int min_pga = (hsl_pga_value<hsr_pga_value? hsl_pga_value:hsr_pga_value);
-	int i = 0;
+	unsigned int i = 0;
 
 	hi64xx_update_bits(codec, HI6402_HPL_POP_CFG1_REG, 0x3F, 0x01F);
 	/* -32dB init */
@@ -4415,7 +4416,7 @@ static void hi6402_dump_reg(char *buf, unsigned int dump_size)
 	len = strlen(buf);
 	len += snprintf(buf + len, dump_size - len,  "%#04x-%#04x\n", HI6402_IRQ_REG_IOMUX,
 		snd_soc_read(g_hi6402_codec, HI6402_IRQ_REG_IOMUX -CODEC_BASE_ADDR));
-	if (len >= dump_size)
+	if (((unsigned int)len) >= dump_size)
 		return;
 
 	for (i = BASE_ADDR_PAGE_CFG + 0x00; i <= BASE_ADDR_PAGE_CFG + 0xff; i++)
@@ -4425,21 +4426,21 @@ static void hi6402_dump_reg(char *buf, unsigned int dump_size)
 			continue;
 		}
 		len += snprintf(buf + len, dump_size - len, "%#04x-%#04x\n", CODEC_BASE_ADDR | i, snd_soc_read(g_hi6402_codec, i));
-		if (len >= dump_size)
+		if (((unsigned int)len) >= dump_size)
 			return;
 	}
 
 	for (i = BASE_ADDR_PAGE_CFG + 0x41; i <= BASE_ADDR_PAGE_CFG + 0x4c; i++)
 	{
 		len += snprintf(buf + len, dump_size - len, "%#04x-%#04x\n", CODEC_BASE_ADDR | i, snd_soc_read(g_hi6402_codec, i));
-		if (len >= dump_size)
+		if (((unsigned int)len) >= dump_size)
 			return;
 	}
 
 	for (i = BASE_ADDR_PAGE_DIG + 0x00; i <= BASE_ADDR_PAGE_DIG + 0x1FF; i++)
 	{
 		len += snprintf(buf + len, dump_size - len, "%#04x-%#04x\n", CODEC_BASE_ADDR | i, snd_soc_read(g_hi6402_codec, i));
-		if (len >= dump_size)
+		if (((unsigned int)len) >= dump_size)
 			return;
 	}
 	len = strlen(buf);
@@ -4921,7 +4922,7 @@ MODULE_DEVICE_TABLE(of, hi6402_platform_match);
 
 static void hi6402_get_board_cfg(struct device_node *node, struct hi6402_board_cfg *board_cfg)
 {
-	int val = 0;
+	unsigned int val = 0;
 
 	/* get board defination */
 	if (!of_property_read_u32(node, "micbias1_mic2", &val)) {
@@ -4931,13 +4932,13 @@ static void hi6402_get_board_cfg(struct device_node *node, struct hi6402_board_c
 	}
 
 	if (!of_property_read_u32(node, "hisilicon,mic_num", &val)) {
-		board_cfg->mic_num = val;
+		board_cfg->mic_num = (int)val;
 	} else {
 		board_cfg->mic_num = 2;
 	}
 
 	if (!of_property_read_u32(node, "hisilicon,hac_gpio", &val)) {
-		board_cfg->hac_gpio = val;
+		board_cfg->hac_gpio = (int)val;
 	} else {
 		board_cfg->hac_gpio = 0;
 	}
@@ -4989,7 +4990,7 @@ static void hi6402_init_codec_device(struct hi_cdc_ctrl* cdc)
 	hi_cdcctrl_reg_write(cdc, HI6402_IOS_IOM_CTRL33, 0x104);
 	hi_cdcctrl_reg_write(cdc, HI6402_IRQ_REG_IRQ_CTRL, 0x84);
 }
-
+/*lint -e429*/
 static int hi6402_platform_probe(struct platform_device *pdev)
 {
 	struct device *dev = &pdev->dev;
@@ -5088,7 +5089,7 @@ free_platform_data:
 
 	return ret;
 }
-
+/*lint +e429*/
 static int hi6402_platform_remove(struct platform_device *pdev)
 {
 	struct hi6402_platform_data *priv =
